python/pyproject.toml
[project]
name = "shinemonitor-api"
version = "0.5.0"
description = "ShineMonitor Rest-API client in Python (WatchPower, SolarPower, etc.)"
authors = [{ name = "David Freire", email = "davidsmfreire@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "httpx>=0.27,<1",
    "pydantic>=1.10,<3",
]

[dependency-groups]
examples = [
    "python-dotenv>=1.0.1,<2.0.0",
    "pandas==1.5.3",
    "tqdm>=4.67.1,<5.0.0",
]
dev = [
    "python-semantic-release>=9.15.2,<10.0.0",
    "pre-commit==4.0.0",
    "mypy>=1.14.1,<2.0.0",
]
test = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "shinemonitor-mock",
]

[tool.uv.sources]
shinemonitor-mock = { path = "../mock-server", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
exclude = ["shinemonitor_api/_methods\\.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shinemonitor_api"]

[tool.semantic_release]
assets = []
version_toml = ["pyproject.toml:project.version"]
version_variables = ["shinemonitor_api/__init__.py:__version__"]
build_command = "pip install uv && uv build"
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = []
mode = "init"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = false

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
other_allowed_tags = ["build", "chore", "ci", "docs", "style", "refactor", "test"]
allowed_tags = ["feat", "fix", "perf"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
